requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# Data
yfinance>=0.2.28

# Technical Analysis
ta>=0.11.0

# ML
scikit-learn>=1.3.0
lightgbm>=4.0.0
optuna>=3.3.0

# Deep Learning
torch>=2.0.0
pytorch-lightning>=2.0.0
darts>=0.27.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0

# Utils
joblib>=1.3.0
tqdm>=4.65.0
pyarrow>=12.0.0
